+++ packages/terra-date-input/tests/wdio/date-input-spec.js
@@ -1,36 +1,36 @@
 Terra.describeViewports('Date Input', ['medium'], () => {
   it('should display no date provided', () => {
-    browser.url('/raw/tests/terra-date-input/date-input/default-date-input');
+    browser.url('/raw/tests/cerner-terra-framework-docs/date-input/default-date-input');
     Terra.validates.element('no date');
   });
 
   it('should display full date provided', () => {
-    browser.url('/raw/tests/terra-date-input/date-input/populated-date-input');
+    browser.url('/raw/tests/cerner-terra-framework-docs/date-input/populated-date-input');
     Terra.validates.element('full date');
   });
 
   it('should display invalid date provided', () => {
-    browser.url('/raw/tests/terra-date-input/date-input/invalid-provided-date-input');
+    browser.url('/raw/tests/cerner-terra-framework-docs/date-input/invalid-provided-date-input');
     Terra.validates.element('invalid date');
   });
 
   it('should be invalid', () => {
-    browser.url('/raw/tests/terra-date-input/date-input/invalid-date-input');
+    browser.url('/raw/tests/cerner-terra-framework-docs/date-input/invalid-date-input');
     Terra.validates.element('invalid input');
   });
 
   it('should display incomplete input', () => {
-    browser.url('/raw/tests/terra-date-input/date-input/incomplete-date-input');
+    browser.url('/raw/tests/cerner-terra-framework-docs/date-input/incomplete-date-input');
     Terra.validates.element('incomplete');
   });
 
   it('should display invalid and incomplete date input', () => {
-    browser.url('/raw/tests/terra-date-input/date-input/invalid-incomplete-date-input');
+    browser.url('/raw/tests/cerner-terra-framework-docs/date-input/invalid-incomplete-date-input');
     Terra.validates.element('invalid and incomplete');
   });
 
   it('should display Disabled Date Input', () => {
-    browser.url('/raw/tests/terra-date-input/date-input/disabled-date-input');
+    browser.url('/raw/tests/cerner-terra-framework-docs/date-input/disabled-date-input');
     browser.refresh();
     // Removes the blinking cursor to prevent screenshot mismatches.
     Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -39,7 +39,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
   });
 
   it('should display Day Month Year Date Input', () => {
-    browser.url('/raw/tests/terra-date-input/date-input/day-month-year-date-input');
+    browser.url('/raw/tests/cerner-terra-framework-docs/date-input/day-month-year-date-input');
     browser.refresh();
     // Removes the blinking cursor to prevent screenshot mismatches.
     Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -48,7 +48,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
   });
 
   it('should display Month Day Year Date Input', () => {
-    browser.url('/raw/tests/terra-date-input/date-input/month-day-year-date-input');
+    browser.url('/raw/tests/cerner-terra-framework-docs/date-input/month-day-year-date-input');
     browser.refresh();
     // Removes the blinking cursor to prevent screenshot mismatches.
     Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -57,26 +57,26 @@ Terra.describeViewports('Date Input', ['medium'], () => {
   });
 
   it(' should Focus Month Styles', () => {
-    browser.url('/raw/tests/terra-date-input/date-input/default-date-input');
+    browser.url('/raw/tests/cerner-terra-framework-docs/date-input/default-date-input');
     $('select[name="terra-date-month-date-input"]').click();
     Terra.validates.element('focus month styles');
   });
 
   it('should Focus Day Styles', () => {
-    browser.url('/raw/tests/terra-date-input/date-input/populated-date-input');
+    browser.url('/raw/tests/cerner-terra-framework-docs/date-input/populated-date-input');
     $('input[name="terra-date-day-date-input"]').click();
     Terra.validates.element('focus day styles');
   });
 
   it('should display Focus Year Styles', () => {
-    browser.url('/raw/tests/terra-date-input/date-input/populated-date-input');
+    browser.url('/raw/tests/cerner-terra-framework-docs/date-input/populated-date-input');
     $('input[name="terra-date-year-date-input"]').click();
     Terra.validates.element('focus year styles');
   });
 
   describe('Month Select placeholder option clears value', () => {
     it('Selects January in month select', () => {
-      browser.url('/raw/tests/terra-date-input/date-input/day-month-year-date-input');
+      browser.url('/raw/tests/cerner-terra-framework-docs/date-input/day-month-year-date-input');
       Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
       Terra.hideInputCaret('input[name="terra-date-year-date-input"]');
       $('select[name="terra-date-month-date-input"]').click();
@@ -96,7 +96,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
   describe('Month select keyboard operations', () => {
     describe('Month Select UP_ARROW increments month by 1', () => {
       it('should increment month by 1', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/populated-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/populated-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -112,7 +112,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
 
     describe('Month Select DOWN_ARROW decrements month by 1', () => {
       it('should decrement month by 1', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/populated-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/populated-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -127,7 +127,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
 
     describe('Month Select Pressing DELETE in month with value clears the value', () => {
       it('should clear value on delete', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/populated-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/populated-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -142,7 +142,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
 
     describe('Month Select Pressing DELETE in month with no value shifts focus to day in d-m-y format', () => {
       it('should shift focus to day', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/day-month-year-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/day-month-year-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -158,7 +158,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
 
     describe('Month Select Pressing DELETE in month with no value does not shift focus to day in m-d-y format', () => {
       it('should not shift focus to day', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/month-day-year-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/month-day-year-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -176,7 +176,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
   describe('Day input keyboard operations', () => {
     describe('Day Input UP_ARROW increments day by 1 when day value is filled in', () => {
       it('should increment day by 1', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/populated-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/populated-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -191,7 +191,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
 
     describe('Day Input UP_ARROW increments day by 1 when day value is empty', () => {
       it('should increment day by 1', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/default-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/default-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -206,7 +206,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
 
     describe('Day Input UP_ARROW is cycles to 01 when the day has reached 31', () => {
       it('should cycle day to 01', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/default-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/default-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -221,7 +221,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
 
     describe('Day Input DOWN_ARROW decrements day by 1 when day value is filled in', () => {
       it('should decrement day by 1', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/populated-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/populated-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -235,7 +235,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
 
     describe('Day Input DOWN_ARROW decrements day to 31 when day value is empty', () => {
       it('should decrement day to 31', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/default-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/default-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -250,7 +250,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
 
     describe('Day Input DOWN_ARROW cycles to 31 when the day has reached 01', () => {
       it('should cycle day to 31', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/default-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/default-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -265,7 +265,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
 
     describe('Day Input Pressing DELETE in day input with value clears the character', () => {
       it('should clear day value on delete', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/populated-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/populated-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -280,7 +280,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
 
     describe('Day Input Pressing DELETE in day input with no value shifts focus day to month in m-d-y format', () => {
       it('should shift focus to month', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/month-day-year-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/month-day-year-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -294,7 +294,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
 
     describe('Day Input Pressing DELETE in day input with no value does not shift focus in d-m-y format', () => {
       it('should not shift focus on delete', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/day-month-year-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/day-month-year-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -309,7 +309,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
 
     describe('Day Input Pressing T in day input', () => {
       it('should set date to current if blank', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/default-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/default-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -322,7 +322,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
       });
 
       it('should set date to current if valid value is there', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/populated-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/populated-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -335,7 +335,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
       });
 
       it('should set date to current if invalid value is there', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/default-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/default-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -355,7 +355,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
 
     describe('Day Input Pressing - in day input', () => {
       it('should set date to yesterday if blank', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/default-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/default-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -369,7 +369,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
       });
 
       it('should set date to current value - 1 day if valid value is there', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/populated-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/populated-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -382,7 +382,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
       });
 
       it('should set date to yesterday if no value is there', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/default-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/default-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -401,7 +401,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
       });
 
       it('should set date to yesterday if invalid value is there', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/default-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/default-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -423,7 +423,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
 
     describe('Day Input Pressing + in day input', () => {
       it('should set date to tomorrow if blank', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/default-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/default-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -437,7 +437,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
       });
 
       it('should set date to current value + 1 day if valid value is there', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/populated-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/populated-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -450,7 +450,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
       });
 
       it('should set date to tomorrow if no value is there', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/default-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/default-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -469,7 +469,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
       });
 
       it('should set date to tomorrow if invalid value is there', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/default-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/default-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -493,7 +493,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
   describe('Year input keyboard operations', () => {
     describe('Year Input UP_ARROW increments year by 1 when year value is filled in', () => {
       it('should increment year by 1', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/populated-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/populated-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -508,7 +508,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
 
     describe('Year Input UP_ARROW increments year to 1900 when year value is empty', () => {
       it('should increment year to 1900', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/default-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/default-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -523,7 +523,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
 
     describe('Year Input UP_ARROW is cycles to 1900 when the year has reached 2100', () => {
       it('should cycle to 1900', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/default-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/default-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -538,7 +538,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
 
     describe('Year Input DOWN_ARROW decrements year by 1 when year value is filled in', () => {
       it('should decrement year by 1', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/populated-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/populated-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -553,7 +553,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
 
     describe('Year Input DOWN_ARROW decrements year to 2100 when year value is empty', () => {
       it('should decrement year to 2100', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/default-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/default-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -566,7 +566,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
     });
 
     it('should cycle to 2100 when the year has reached 1900', () => {
-      browser.url('/raw/tests/terra-date-input/date-input/default-date-input');
+      browser.url('/raw/tests/cerner-terra-framework-docs/date-input/default-date-input');
       browser.refresh();
       // Removes the blinking cursor to prevent screenshot mismatches.
       Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -577,7 +577,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
     });
 
     it('should not allow user to enter year values outside 1900-2100 ', () => {
-      browser.url('/raw/tests/terra-date-input/date-input/default-date-input');
+      browser.url('/raw/tests/cerner-terra-framework-docs/date-input/default-date-input');
       browser.refresh();
       // Removes the blinking cursor to prevent screenshot mismatches.
       Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -591,7 +591,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
 
     describe('Year Input Pressing DELETE in year input with value clears the character', () => {
       it('should clear year input on delete', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/populated-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/populated-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -606,7 +606,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
 
     describe('Year Input Pressing DELETE in year input with no value shifts focus year to day in m-d-y format', () => {
       it('should shift focus to day', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/month-day-year-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/month-day-year-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -620,7 +620,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
 
     describe('Year Input Pressing DELETE in year input with no value shifts focus year to month in d-m-y format', () => {
       it('should shift focus to month', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/day-month-year-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/day-month-year-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -634,7 +634,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
 
     describe('Year Input Pressing T in year input', () => {
       it('should set date to current if blank', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/default-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/default-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -647,7 +647,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
       });
 
       it('should set date to current if valid value is there', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/populated-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/populated-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -660,7 +660,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
       });
 
       it('should set date to current if invalid value is there', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/default-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/default-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -680,7 +680,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
 
     describe('Year Input Pressing - in year', () => {
       it('should set date to yesterday if blank', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/default-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/default-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -694,7 +694,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
       });
 
       it('should set date to current value - 1 day if valid value is there', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/populated-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/populated-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -707,7 +707,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
       });
 
       it('should set date to yesterday if no value is there', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/default-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/default-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -726,7 +726,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
       });
 
       it('should set date to yesterday if invalid value is there', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/default-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/default-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -748,7 +748,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
 
     describe('Year Input Pressing + in year input', () => {
       it('should set date to tomorrow if blank', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/default-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/default-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -762,7 +762,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
       });
 
       it('should set date to current value + 1 day if valid value is there', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/populated-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/populated-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -775,7 +775,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
       });
 
       it('should set date to tomorrow if no value is there', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/default-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/default-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -794,7 +794,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
       });
 
       it('should set date to tomorrow if invalid value is there', () => {
-        browser.url('/raw/tests/terra-date-input/date-input/default-date-input');
+        browser.url('/raw/tests/cerner-terra-framework-docs/date-input/default-date-input');
         browser.refresh();
         // Removes the blinking cursor to prevent screenshot mismatches.
         Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -817,7 +817,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
 
   describe('Date Input OnChange Valid Date', () => {
     it('should trigger onChange on valid input', () => {
-      browser.url('/raw/tests/terra-date-input/date-input/default-date-input');
+      browser.url('/raw/tests/cerner-terra-framework-docs/date-input/default-date-input');
       browser.refresh();
       // Removes the blinking cursor to prevent screenshot mismatches.
       Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -838,7 +838,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
 
   describe('Date Input OnChange Invalid Date', () => {
     it('should trigger onChange on invalid input', () => {
-      browser.url('/raw/tests/terra-date-input/date-input/default-date-input');
+      browser.url('/raw/tests/cerner-terra-framework-docs/date-input/default-date-input');
       browser.refresh();
       // Removes the blinking cursor to prevent screenshot mismatches.
       Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -858,7 +858,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
 
   describe('Date Input prepends 0 on single digit day onBlur', () => {
     it('should prepend 0 onBlur', () => {
-      browser.url('/raw/tests/terra-date-input/date-input/default-date-input');
+      browser.url('/raw/tests/cerner-terra-framework-docs/date-input/default-date-input');
       browser.refresh();
       // Removes the blinking cursor to prevent screenshot mismatches.
       Terra.hideInputCaret('input[name="terra-date-day-date-input"]');
@@ -874,7 +874,7 @@ Terra.describeViewports('Date Input', ['medium'], () => {
 
   describe('Date Input onBlur', () => {
     before(() => {
-      browser.url('/raw/tests/terra-date-input/date-input/focus-blur-date-input');
+      browser.url('/raw/tests/cerner-terra-framework-docs/date-input/focus-blur-date-input');
       $('select[name="terra-date-month-date-input"]').click();
       // eslint-disable-next-line prefer-arrow-callback
       browser.execute(function dispatch() {