+++ packages/terra-framework-docs/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Migrated docs from terra date input.
+
 ## 1.6.0 - (November 16, 2021)
 
 * Changed